chatbot_backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beanie==1.29.0
beautifulsoup4==4.12.2
boto3==1.36.25
botocore==1.36.25
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.2.0
colorama==0.4.6
coverage==7.8.2
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss-cpu==1.11.0
fastapi==0.115.2
fastui==0.7.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.31.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-aws==0.2.13
langchain-community==0.3.18
langchain-core==0.3.61
langchain-mcp-adapters==0.1.7
langchain-ollama==0.3.3
langchain-openai==0.3.6
langchain-text-splitters==0.3.8
langgraph==0.2.74
langgraph-checkpoint==2.0.26
langgraph-checkpoint-mongodb==0.1.0
langgraph-sdk==0.1.70
langsmith==0.3.42
lazy-model==0.2.0
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.4 
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
ollama==0.4.8
openai==1.58.1
openpyxl==3.1.5
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.6.0
propcache==0.3.1
protobuf==6.31.0
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.3
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
Pygments==2.19.2
pymongo==4.9.2
PyPDF2==3.0.1
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
s3transfer==0.11.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentry-sdk==2.19.2
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
sse-starlette==2.3.6
starlette==0.40.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.32.0
watchfiles==1.0.5
websockets==15.0.1
XlsxWriter==3.2.3
yarl==1.20.0
zstandard==0.23.0
unidecode==1.3.7
nltk==3.8.1
chromadb==1.0.10